+++ lib/es_tree/template_element.ex
@@ -2,11 +2,11 @@ defmodule ESTree.TemplateElement do
   @type t :: %ESTree.TemplateElement{ 
     type: binary, 
     loc: ESTree.SourceLocation.t | nil,
-    value: %{cooked: binary, value: binary},
+    value: %{cooked: binary, raw: binary},
     tail: boolean
   }
   defstruct type: "TemplateElement", 
             loc: nil,
-            value: %{cooked: "", value: ""},
+            value: %{cooked: "", raw: ""},
             tail: false
 end